pyproject.toml
[tool.poetry]
name = "ury-wrapped"
version = "0.1.0"
description = ""
authors = ["Marks Polakovs <marks@markspolakovs.me>"]
readme = "README.md"
packages = [{include = "ury_wrapped"}]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2 = "^2.9.5"
PyYAML = "^6.0"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
Pillow = "^9.3.0"
Flask = "^2.2.2"
requests = "^2.28.1"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
